% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc_dynamic_require.R
\name{bioc_dynamic_require}
\alias{bioc_dynamic_require}
\title{Load or install a package available at Biocinductor}
\usage{
bioc_dynamic_require(package)
}
\arguments{
\item{package}{character vector with name of the package you need to load/install.}
}
\value{
The function loads the package if you already have it installed, or installs from bioconductor if it is not installed yet.
It might ask for permission to update libraries when installing.
}
\description{
Load or install a package available at Biocinductor
}
\examples{

bioc_dynamic_require("edgeR")

}
