% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_require.R
\name{dynamic_require}
\alias{dynamic_require}
\title{Load or install a package available at CRAN}
\usage{
dynamic_require(package)
}
\arguments{
\item{package}{Name of the package between quotes (character vector).}
}
\value{
Loading of the package or installing of it. It might ask for permission to update libraries when installing.
}
\description{
Load or install a package available at CRAN
}
\examples{
dynamic_require("usethis")

x <- "wesanderson"
dynamic_require("wesanderson")

}
