% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_dynamic_require.R
\name{github_dynamic_require}
\alias{github_dynamic_require}
\title{Load or install a package available at github}
\usage{
github_dynamic_require(package, repository)
}
\arguments{
\item{package}{Name of the package between quotes (character vector).}

\item{repository}{Repository address in the format "user/package", between quotes (character vector)}
}
\value{
Loading of the package or installing of it. It might ask for permission to update libraries when installing.
}
\description{
Load or install a package available at github
}
\examples{

github_dynamic_require("ggpubr", "kassambara/ggpubr")

}
